split relation utils
