+++ src/test/scala/cognite/spark/v1/fdm/FlexibleDataModelRelationUtilsTest.scala
@@ -1,6 +1,6 @@
 package cognite.spark.v1.fdm
 
-import cognite.spark.v1.fdm.FlexibleDataModelRelationUtils._
+import cognite.spark.v1.fdm.RelationUtils.FlexibleDataModelRelationUtils._
 import cognite.spark.v1.fdm.utils.FDMViewPropertyDefinitions._
 import com.cognite.sdk.scala.v1.fdm.common.DirectRelationReference
 import com.cognite.sdk.scala.v1.fdm.instances.InstancePropertyValue.{ViewDirectNodeRelation, ViewDirectNodeRelationList}
